Adds auditing.
Adds `audits` table that will track whether a given issue is marked to
be kept or closed by a user.

Currently this is used to filter issues that were flagged to keep out of
the list of potential issues to show. This was made configurable on the
back end but no via the interface yet.

* Creates `Audit` record when keeping or closing issue
* Creates `IssueFinder` module that handles finding issues and scrubbing
against issues that have been flagged to keep or close

This also sets up Ecto Timestamps to use `Timex`. This allows for the
use of `Timex` helpers (such as `from_now`).