Adds the ability to view already audited issues.
Sets up a plug on the issue controller that checks if users want to look
at all issues (this would include issues that were flagged to be kept
open) or only issues that haven't been "audited" already.

The templates should be cleaned up to no longer include logic that
checks this.