Changes where repo list link takes you.
Clicking on a repo in the list of repos will now take you directly to a
random issue. Previously it would take you to `Issue#index` which have
the issue count and a button to start cleaning. Now that we have the
issue count in the list this page felt less useful. The index page was
also unstyled so this saved on that too.

I decided to leave `Issue#index` around for now since I'm not sure
exactly how I want the user flow to look. The issue index also accounted
for issues that were flagged to be kept which the repo list does not.