Changes how users interact with issues.
* Uses `Swing` JS library to allow users to swipe issues to keep or
close
* Instead of using `IssueController#show` now uses
`IssueController#index` to list all issues and have users swipe through
the entire stack
* Uses Channels to communicate with the server to keep and close issues

Some notes on implementation:

* Initially I tried to make every `issue` DOM element a "card" on the
stack. This had bad performance on repos that had a large number of
issues. I decided to only have one "card" at a time and, when an action
is made, convert the next issue into a "card". We should see if this
causes issues in the future where maybe the conversion into a card is
slow.
* Similar to the above issue I initially showed all card. Even with CSS
to try to make the issues all the same size when showing all issues it
would look like there was a border because of so many stacked cards. I
decided to start with all issues hidden, initially show 5 then show an
issue each time the top issue was kept or closed.

TODO:

* Socket token salt should probably be an environmental variabl
* Currently all the `Show` stuff was left it - we need to decide what we
want to do with it
* I would like to update the URL to be based on the current issue so
links can be shared - this ties in to the point above.